requirements.txt

fastapi
uvicorn
python-multipart
httpx
python-dotenv
openai
playsound
google-search-results
langchain
langgraph
spoon-ai-sdk
